Reorganized commands of similar categories into their own packages
